Add files via upload
Here's a Hangman Game. Hangman is a paper and pencil guessing game for two or more players. One player thinks of a word, phrase or sentence and the other tries to guess it by suggesting letters within a certain number of guesses. Lets see who wins the game. All The Best !!.